requirements.txt
anthropic==0.49.0
anyio==3.7.1
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.83.0
filelock==3.18.0
fsspec==2025.3.0
h11==0.14.0
hatesonar==0.0.5
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
jiter==0.9.0
joblib==1.4.2
nltk==3.6.7
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pdfminer.six==20211012
pdfplumber==0.6.0
pillow==11.1.0
pycparser==2.22
pydantic==1.10.21
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.54
starlette==0.19.1
textblob==0.15.3
threadpoolctl==3.6.0
tokenizers==0.21.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.17.0
vaderSentiment==3.3.2
Wand==0.6.13
